Adds account type to signup pane
Also decomposes the GUI components into separate classes
